get total processes on current system
In order to get the total processes (which itself isn't terribly
difficult), we need to pull in the `sysinfo` crate which does all the
low-level kernel fetching for us.

The problem here is that `sysinfo::System` needs to be kept around, so
we're going to have to initialise that as high up as possible (currently
being done in `main.rs`), so that the rest of the application can access
it.

I think this is going to need a bit of a move in the future, but at
least for right now we're able to pass it down to the UI layer and have
that get the information it needs.

Annoying that `Application::connect_activate()` requires ownership of
all the values it's passed, which means we have to use a rather nasty
`move` call to appease the compiler.